2024-04-25 General 수정 보완

+++ src/main/java/com/winter/app/general/GeneralService.java
@@ -18,36 +18,40 @@ public class GeneralService {
         return generalDAO.getManageList();
     }
 
-    public int updateCarManage(CarManageVO carManageVO) throws Exception{
+    public int updateCarManage(CarManageVO carManageVO) throws Exception {
         return generalDAO.updateCarManage(carManageVO);
     }
 
-    public int deleteCarManage(CarManageVO carManageVO) throws Exception{
-        return generalDAO.deleteCarManage(carManageVO);
+    public int deleteCarManage(CarManageVO carManageVO) throws Exception {
+        CarDetailVO carDetailVO = new CarDetailVO();
+        carDetailVO.setCar_code(carManageVO.getCar_code());
+        carDetailVO.setPro_status(0L);
+        int result = generalDAO.updateStatus(carDetailVO);
+        if (result == 1) {
+            result = generalDAO.deleteCarManage(carManageVO);
+        }else {
+            return result;
+        }
+        return result;
     }
 
-    public List<CarDetailVO> getUsableList() throws Exception{
-        return generalDAO.getUsableList();
-    }
-
-    public CarManageVO carAllocation(CarManageVO carManageVO) throws Exception{
+    public CarManageVO carAllocation(CarManageVO carManageVO) throws Exception {
 
         //배차 요청 정보 등록
         int result = generalDAO.carAllocation(carManageVO);
-        CarDetailVO carDetailVO = new CarDetailVO();
-        // 해당 차량 사용 가능 상태에서 배차요청 상태로 변경
-        carDetailVO.setCar_code(carManageVO.getCar_code());
-
-        // 해당 현장의 배차코드 넘버 부여
-        generalDAO.updateStatus1(carDetailVO);
-
+        if (result == 1) {
+            CarDetailVO carDetailVO = new CarDetailVO();
+            // 해당 차량 사용 가능 상태에서 배차요청 상태로 변경
+            carDetailVO.setCar_code(carManageVO.getCar_code());
+            carDetailVO.setPro_status(1L);
+            // 해당 현장의 배차코드 넘버 부여
+            generalDAO.updateStatus(carDetailVO);
+        }
         carManageVO = generalDAO.getInfo(carManageVO);
-
-        System.out.println("==========\n" + carManageVO);
-        return  carManageVO;
+        return carManageVO;
     }
 
-    public CarDetailVO getCarNumber (CarDetailVO carDetailVO) throws Exception{
+    public CarDetailVO getCarNumber(CarDetailVO carDetailVO) throws Exception {
 
         return generalDAO.getCarNumber(carDetailVO);
     }
